more convenience (#146)
* more convenience

* usize

* creder

* format

* basic creder tests

* run json through jq

* better coverage

* more tests

* makefile improvement

* fix test

* fix tests more

* version bump

+++ src/core/counter/mod.rs
@@ -303,8 +303,8 @@ impl Counter {
         Ok(())
     }
 
-    pub fn full_size(&self) -> Result<u32> {
-        Ok(tables::sizage(&self.code())?.fs)
+    pub fn full_size(&self) -> Result<usize> {
+        Ok(tables::sizage(&self.code())?.fs as usize)
     }
 }
 
@@ -537,7 +537,7 @@ mod test {
     #[case("-AAB", counter::Codex::ControllerIdxSigs, 4)]
     #[case("-0VAAAQA", counter::Codex::BigAttachedMaterialQuadlets, 8)]
     #[case("--AAAAAA", counter::Codex::KERIProtocolStack, 8)]
-    fn qb_size(#[case] qsc: &str, #[case] code: &str, #[case] full_size: u32) {
+    fn qb_size(#[case] qsc: &str, #[case] code: &str, #[case] full_size: usize) {
         let counter = Counter::new(None, None, None, None, Some(qsc), None).unwrap();
         assert_eq!(counter.code(), code); // Just a self-check of the input data
         assert_eq!(counter.full_size().unwrap(), full_size);
